pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [

]

[project.optional-dependencies]
dev = [
    "aws-cdk-lib",
    "constructs",
    "aws-cdk.aws-iam",
    # Add any other development tools here, e.g., "pytest", "black"
]

# [build-system]
# requires = ["uv"]
# build-backend = "uv.build"
